+++ fetch_data/api_data/time_to_first_value/create_time_to_first_value_observation.py
@@ -1,13 +1,21 @@
 """Create Time to First Value Observation"""
 
 from datetime import datetime
-from typing import Iterable, Optional
+from typing import Optional
 from requests.exceptions import HTTPError, ReadTimeout, SSLError, TooManyRedirects
 from urllib3.exceptions import ConnectTimeoutError, MaxRetryError, NameResolutionError
 
 from data import get_async_session
 from fetch_data.api_data.log_data.fetch_log_data import get_log_param_string
-from fetch_data.utils import dict_to_url, fetch_api_data, get_wikibase_from_database
+from fetch_data.api_data.time_to_first_value.get_deleted_revisions_data import (
+    get_deleted_item_q_creation_date,
+    get_deleted_q_creation_date,
+)
+from fetch_data.api_data.time_to_first_value.get_revisions_data import (
+    get_item_q_creation_date,
+    get_q_creation_date,
+)
+from fetch_data.utils import fetch_api_data, get_wikibase_from_database
 from logger import logger
 from model.database import (
     WikibaseModel,
@@ -103,156 +111,6 @@ async def get_item_range_creation_date(
             return WikibaseItemDateModel(
                 item_number=i, creation_date=item_creation_date
             )
-    return None
-
-
-async def get_q_creation_date(
-    wikibase: WikibaseModel, item_number: int
-) -> Optional[datetime]:
-    """Get Q# Creation Date"""
-
-    try:
-        rev_result = await fetch_api_data(
-            wikibase.action_api_url()
-            + get_revision_param_string(titles=[f"Q{item_number}"], prop=["timestamp"])
-        )
-        return parse_revision_timestamp(rev_result)
-    except HTTPError:
-        return None
-
-
-async def get_item_q_creation_date(
-    wikibase: WikibaseModel, item_number: int
-) -> Optional[datetime]:
-    """Get Item:Q# Creation Date"""
-
-    try:
-        rev_result = await fetch_api_data(
-            wikibase.action_api_url()
-            + get_revision_param_string(
-                titles=[f"Item:Q{item_number}"], prop=["timestamp"]
-            )
-        )
-        return parse_revision_timestamp(rev_result)
-    except HTTPError:
-        return None
-
-
-def get_revision_param_string(
-    titles: list[str],
-    limit: int = 1,
-    oldest: bool = True,
-    prop: Optional[list[str]] = None,
-) -> str:
-    """Revision URL Parameters"""
-
-    parameters: dict = {
-        "action": "query",
-        "format": "json",
-        "prop": "revisions",
-        "titles": "|".join(titles),
-        "rvdir": "newer" if oldest else "older",
-        # "formatversion": 2,
-        "rvlimit": limit,
-    }
-    if prop is not None:
-        parameters["rvprop"] = "|".join(prop)
-    return dict_to_url(parameters)
-
-
-def parse_revision_timestamp(revision_result: dict) -> Optional[datetime]:
-    """Parse Timestamp from Revision"""
-
-    try:
-        rev_page_id_list: Iterable[str] = revision_result["query"]["pages"].keys()
-        assert len(rev_page_id_list) > 0
-        for id in rev_page_id_list:
-            if "revisions" in revision_result["query"]["pages"][id]:
-                item_creation_date = datetime.strptime(
-                    revision_result["query"]["pages"][id]["revisions"][0][
-                        "timestamp"
-                    ],
-                    "%Y-%m-%dT%H:%M:%SZ",
-                )
-                return item_creation_date
-    except KeyError as exc:
-        logger.debug(revision_result)
-        raise exc
-    return None
-
-
-async def get_deleted_q_creation_date(
-    wikibase: WikibaseModel, item_number: int
-) -> Optional[datetime]:
-    """Get Deleted Q# Creation Date"""
-
-    try:
-        del_rev_result = await fetch_api_data(
-            wikibase.action_api_url()
-            + get_del_rev_param_string(titles=[f"Q{item_number}"], prop=["timestamp"])
-        )
-        return parse_del_rev_timestamp(del_rev_result)
-    except HTTPError:
-        return None
-
-
-async def get_deleted_item_q_creation_date(
-    wikibase: WikibaseModel, item_number: int
-) -> Optional[datetime]:
-    """Get Deleted Item:Q# Creation Date"""
-
-    try:
-        del_rev_result = await fetch_api_data(
-            wikibase.action_api_url()
-            + get_del_rev_param_string(
-                titles=[f"Item:Q{item_number}"], prop=["timestamp"]
-            )
-        )
-        return parse_del_rev_timestamp(del_rev_result)
-    except HTTPError:
-        return None
-
-
-def get_del_rev_param_string(
-    titles: list[str],
-    limit: int = 1,
-    oldest: bool = True,
-    prop: Optional[list[str]] = None,
-) -> str:
-    """Deleted Revision URL Parameters"""
-
-    parameters: dict = {
-        "action": "query",
-        "format": "json",
-        "prop": "deletedrevisions",
-        "titles": "|".join(titles),
-        "drvdir": "newer" if oldest else "older",
-        # "formatversion": 2,
-        "drvlimit": limit,
-    }
-    if prop is not None:
-        parameters["drvprop"] = "|".join(prop)
-    return dict_to_url(parameters)
-
-
-def parse_del_rev_timestamp(del_rev_result: dict) -> Optional[datetime]:
-    """Parse Timestamp from Deleted Revision"""
-
-    try:
-        del_rev_page_id_list: Iterable[str] = del_rev_result["query"]["pages"].keys()
-        assert len(del_rev_page_id_list) > 0
-        for id in del_rev_page_id_list:
-            if "deletedrevisions" in del_rev_result["query"]["pages"][id]:
-                item_creation_date = datetime.strptime(
-                    del_rev_result["query"]["pages"][id]["deletedrevisions"][0][
-                        "timestamp"
-                    ],
-                    "%Y-%m-%dT%H:%M:%SZ",
-                )
-                return item_creation_date
-    except KeyError as exc:
-        logger.debug(del_rev_result)
-        raise exc
     return None
